+++ rawbuilder/__init__.py
@@ -2,6 +2,6 @@
 
 __author__ = """Mina Farag Amin"""
 __email__ = 'mina.farag@icloud.com'
-__version__ = '0.0.6'
+__version__ = '0.0.7'
 
 from .dataset.DataSet import DataSet